requirements.txt
asgiref==3.7.2
cloudinary==1.33.0
Django==3.2.19
django-cloudinary-storage==0.3.0
djangorestframework==3.14.0
Pillow==8.2.0
pytz==2023.3
sqlparse==0.4.4
urllib3==1.26.16
